+++ src/components/effects/ScoreMessageEffect.tsx
